Convert --define rules to targets
Remove stratum_bcm_package
