+++ .circleci/config.yml
@@ -202,8 +202,8 @@ jobs:
       - run:
           name: Build stratum_bcm
           command: |
-            bazel build --define bcm_sdk=lt //stratum/hal/bin/bcm/standalone:stratum_bcm_deb
-            cp bazel-bin/stratum/hal/bin/bcm/standalone/stratum_bcm_deb.deb $DOCKER_SCOPE
+            bazel build //stratum/hal/bin/bcm/standalone:stratum_bcm_sdklt_deb
+            cp bazel-bin/stratum/hal/bin/bcm/standalone/stratum_bcm_sdklt_deb.deb $DOCKER_SCOPE/stratum_bcm_deb.deb
       - *docker_login
       - *docker_build
       - *docker_push
@@ -227,8 +227,8 @@ jobs:
           name: Build stratum_bcm
           command: |
             git config --global --unset url.ssh://git@github.com.insteadof
-            bazel build --define bcm_sdk=nsa //stratum/hal/bin/bcm/standalone:stratum_bcm_deb
-            cp bazel-bin/stratum/hal/bin/bcm/standalone/stratum_bcm_deb.deb $DOCKER_SCOPE
+            bazel build //stratum/hal/bin/bcm/standalone:stratum_bcm_opennsa_deb
+            cp bazel-bin/stratum/hal/bin/bcm/standalone/stratum_bcm_opennsa_deb.deb $DOCKER_SCOPE/stratum_bcm_deb.deb
       - *docker_login
       - *docker_build
       - *docker_push
