--- stratum/hal/bin/bcm/standalone/sdklt/load-kernel-modules.sh
@@ -1,30 +1,0 @@
-#!/bin/bash
-#
-# Copyright 2020-present Open Networking Foundation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-set -ex
-
-# Remove other SDK kernel modules, if present
-rmmod linux_bcm_knet || true
-rmmod linux_user_bde || true
-rmmod linux_kernel_bde || true
-
-# Reinsert SDKLT kernel modules
-rmmod linux_ngknet || true
-rmmod linux_ngbde || true
-pushd /usr/lib/stratum/
-insmod linux_ngbde.ko && insmod linux_ngknet.ko
-popd
-sleep 1
