+++ stratum/hal/bin/bcm/standalone/sdklt/start-stratum.sh
@@ -17,18 +17,29 @@
 
 set -ex
 
-load-kernel-modules.sh
+# Remove other SDK kernel modules, if present
+rmmod linux_bcm_knet || true
+rmmod linux_user_bde || true
+rmmod linux_kernel_bde || true
+
+# Reinsert SDKLT kernel modules
+rmmod linux_ngknet || true
+rmmod linux_ngbde || true
+pushd /usr/lib/stratum/
+insmod linux_ngbde.ko && insmod linux_ngknet.ko
+popd
+sleep 1
 
 mkdir -p /var/run/stratum /var/log/stratum
 
 PLATFORM=$(cat /etc/onl/platform)
 
-exec stratum_bcm \
+exec stratum_bcm_sdklt \
     -external_stratum_urls=0.0.0.0:28000 \
     -persistent_config_dir=/etc/stratum \
     -base_bcm_chassis_map_file=/etc/stratum/stratum_configs/${PLATFORM}/base_bcm_chassis_map.pb.txt \
     -chassis_config_file=/etc/stratum/stratum_configs/${PLATFORM}/chassis_config.pb.txt \
-    -bcm_sdk_config_file=/etc/stratum/stratum_configs/${PLATFORM}/SDKLT.yml \
+    -bcm_sdk_config_file=/etc/stratum/stratum_configs/${PLATFORM}/sdklt.yml \
     -phal_config_file=/etc/stratum/stratum_configs/${PLATFORM}/phal_config.pb.txt \
     -bcm_serdes_db_proto_file=/etc/stratum/stratum_configs/dummy_serdes_db.pb.txt \
     -bcm_hardware_specs_file=/etc/stratum/stratum_configs/bcm_hardware_specs.pb.txt \
