--- stratum/hal/lib/bcm/sdk/macros_include.h
@@ -1,34 +1,0 @@
-/*
- * Copyright 2020-present Open Networking Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
- * The Broadcom Switch API header code upon which this file depends is Copyright 2007-2020 Broadcom Inc.
- *
- * This file depends on Broadcom's OpenNSA SDK.
- * Additional license terms for OpenNSA are available from Broadcom or online:
- *     https://github.com/Broadcom-Network-Switching-Software/OpenNSA
- */
-
-#ifndef STRATUM_HAL_LIB_BCM_SDK_MACROS_INCLUDE_H_
-#define STRATUM_HAL_LIB_BCM_SDK_MACROS_INCLUDE_H_
-
-extern "C" {
-#include "stratum/hal/lib/bcm/sdk_build_undef.h"  // NOLINT
-#include "bcm/error.h"
-#include "stratum/hal/lib/bcm/sdk_build_undef.h"  // NOLINT
-}
-
-#endif  // STRATUM_HAL_LIB_BCM_SDK_MACROS_INCLUDE_H_
